feat(adapter): add label view type for various adapters

+++ music/src/main/java/app/simple/felicity/adapters/ui/lists/AdapterFolderHierarchy.kt
@@ -48,7 +48,7 @@ class AdapterFolderHierarchy(contents: FolderHierarchyContents) : RecyclerView.A
         return if (position < folderCount) {
             folders[position].id
         } else {
-            songs[position - folderCount].id.toLong()
+            songs[position - folderCount].id
         }
     }
